changes for .tsconfig; importing .json file for string data; dir/file reorganization
