+++ script.js
@@ -1,4 +1,4 @@
-import { Main } from './dist/Main.js'
+import { Main } from './dist/src/Main.js'
 
 window.onload = function() {
 	new Main().runTime()
