+++ src/classes/StringState.ts
