+++ src/classes/StringStateCollection.ts
