+++ src/classes/StringTension.ts
@@ -49,7 +49,6 @@ class StringTension {
     public makeStringTable(tableId: string, numberId: string) {
         let numStrings = (<HTMLInputElement>document.getElementById(numberId)).value
 
-        console.log(numStrings)
         this.stringTable.setNumStrings(Number(numStrings))
         this.redrawStringTable(tableId)
     }
