+++ src/interfaces/StringMake.ts
@@ -1,3 +1,5 @@
+export { StringMake }
+
 interface StringMake {
    brand: string,
    type: string
