+++ src/static/Utilities.ts
